+++ packages/core/src/components.d.ts
@@ -20,8 +20,8 @@ import { CardClickEvent } from "./components/card/card.interfaces";
 import { CardContainerMode } from "./components/card-container/card-container.interfaces";
 import { DatePickerBlurEvent, DatePickerChangeEvent, DatePickerFocusEvent, DatePickerKeyboardEvent } from "./components/date-picker/date-picker.interfaces";
 import { DocumentCardClickEvent } from "./components/document-card/document-card.interfaces";
-import { DocumentComponentAnnotationsWijzigactie, DocumentComponentInputType, DocumentComponentMarkFunction, DocumentComponentMarkItemHighlightEvent, DocumentComponentMode, DocumentComponentOpenToggleEvent, DocumentComponentOzonContentAnchorClickEvent, DocumentComponentRecursiveToggleEvent, DocumentComponentRecursiveToggleState, DocumentComponentTableOfContentsClickEvent, DocumentComponentToggleAnnotationEvent, DocumentComponentWijzigactie } from "./components/document-component/document-component.interfaces";
-import { OzonContentAnchorClickEvent, OzonContentClickEvent, OzonContentInputType, OzonContentMarkFunction, OzonContentMarkItemHighlightEvent, OzonContentUrlResolver } from "./components/ozon-content/ozon-content.interfaces";
+import { DocumentComponentAnnotationsWijzigactie, DocumentComponentInputType, DocumentComponentMarkFunction, DocumentComponentMarkItemHighlightEvent, DocumentComponentMode, DocumentComponentOpenToggleEvent, DocumentComponentOzonContentClickEvent, DocumentComponentRecursiveToggleEvent, DocumentComponentRecursiveToggleState, DocumentComponentTableOfContentsClickEvent, DocumentComponentToggleAnnotationEvent, DocumentComponentWijzigactie } from "./components/document-component/document-component.interfaces";
+import { OzonContentBegripResolver, OzonContentClickEvent, OzonContentInputType, OzonContentMarkFunction, OzonContentMarkItemHighlightEvent, OzonContentUrlResolver } from "./components/ozon-content/ozon-content.interfaces";
 import { ExpandableAnimationEndEvent, ExpandableAnimationStartEvent } from "./components/expandable/expandable";
 import { HeaderAuthStatus, HeaderCompactMode, HeaderEvent, HeaderMenuItem } from "./components/header/header.interfaces";
 import { HistoryItemClickEvent, HistoryItemType } from "./components/history-item/history-item.interfaces";
@@ -51,7 +51,7 @@ import { TabsSwitchEvent } from "./components/tabs/tabs.interfaces";
 import { Placement } from "@popperjs/core";
 import { TreeViewItem, TreeViewPointerEvent } from "./components/tree-view/tree-view.interfaces";
 import { ViewerGridActiveTabSwitchEvent, ViewerGridChangeSizeAnimationEndEvent, ViewerGridChangeSizeEvent, ViewerGridCloseFilterPanelEvent, ViewerGridCloseOverlayEvent, ViewerGridMainToggleEvent, ViewerGridPanelSize, ViewerGridTab } from "./components/viewer-grid/viewer-grid.interfaces";
-import { OzonContentUrlResolver as OzonContentUrlResolver1 } from "./components";
+import { OzonContentBegripResolver as OzonContentBegripResolver1, OzonContentUrlResolver as OzonContentUrlResolver1 } from "./components";
 export { AccordionInternalState, AccordionVariant } from "./components/accordion/accordion.interfaces";
 export { RenvooiMarkFunction, RenvooiMarkItemHighlightEvent, RenvooiValue } from "./components/renvooi/renvooi.interfaces";
 export { AccordionHeading, AccordionSectionActiveChangeEvent, AccordionSectionAnimationEndEvent, AccordionSectionAnimationStartEvent, AccordionSectionState, AccordionSectionToggleClickEvent, AccordionSectionWijzigactie } from "./components/accordion/components/accordion-section.interfaces";
@@ -67,8 +67,8 @@ export { CardClickEvent } from "./components/card/card.interfaces";
 export { CardContainerMode } from "./components/card-container/card-container.interfaces";
 export { DatePickerBlurEvent, DatePickerChangeEvent, DatePickerFocusEvent, DatePickerKeyboardEvent } from "./components/date-picker/date-picker.interfaces";
 export { DocumentCardClickEvent } from "./components/document-card/document-card.interfaces";
-export { DocumentComponentAnnotationsWijzigactie, DocumentComponentInputType, DocumentComponentMarkFunction, DocumentComponentMarkItemHighlightEvent, DocumentComponentMode, DocumentComponentOpenToggleEvent, DocumentComponentOzonContentAnchorClickEvent, DocumentComponentRecursiveToggleEvent, DocumentComponentRecursiveToggleState, DocumentComponentTableOfContentsClickEvent, DocumentComponentToggleAnnotationEvent, DocumentComponentWijzigactie } from "./components/document-component/document-component.interfaces";
-export { OzonContentAnchorClickEvent, OzonContentClickEvent, OzonContentInputType, OzonContentMarkFunction, OzonContentMarkItemHighlightEvent, OzonContentUrlResolver } from "./components/ozon-content/ozon-content.interfaces";
+export { DocumentComponentAnnotationsWijzigactie, DocumentComponentInputType, DocumentComponentMarkFunction, DocumentComponentMarkItemHighlightEvent, DocumentComponentMode, DocumentComponentOpenToggleEvent, DocumentComponentOzonContentClickEvent, DocumentComponentRecursiveToggleEvent, DocumentComponentRecursiveToggleState, DocumentComponentTableOfContentsClickEvent, DocumentComponentToggleAnnotationEvent, DocumentComponentWijzigactie } from "./components/document-component/document-component.interfaces";
+export { OzonContentBegripResolver, OzonContentClickEvent, OzonContentInputType, OzonContentMarkFunction, OzonContentMarkItemHighlightEvent, OzonContentUrlResolver } from "./components/ozon-content/ozon-content.interfaces";
 export { ExpandableAnimationEndEvent, ExpandableAnimationStartEvent } from "./components/expandable/expandable";
 export { HeaderAuthStatus, HeaderCompactMode, HeaderEvent, HeaderMenuItem } from "./components/header/header.interfaces";
 export { HistoryItemClickEvent, HistoryItemType } from "./components/history-item/history-item.interfaces";
@@ -98,7 +98,7 @@ export { TabsSwitchEvent } from "./components/tabs/tabs.interfaces";
 export { Placement } from "@popperjs/core";
 export { TreeViewItem, TreeViewPointerEvent } from "./components/tree-view/tree-view.interfaces";
 export { ViewerGridActiveTabSwitchEvent, ViewerGridChangeSizeAnimationEndEvent, ViewerGridChangeSizeEvent, ViewerGridCloseFilterPanelEvent, ViewerGridCloseOverlayEvent, ViewerGridMainToggleEvent, ViewerGridPanelSize, ViewerGridTab } from "./components/viewer-grid/viewer-grid.interfaces";
-export { OzonContentUrlResolver as OzonContentUrlResolver1 } from "./components";
+export { OzonContentBegripResolver as OzonContentBegripResolver1, OzonContentUrlResolver as OzonContentUrlResolver1 } from "./components";
 export namespace Components {
     interface DsoAccordion {
         "_getState": () => Promise<AccordionInternalState>;
@@ -580,6 +580,11 @@ export namespace Components {
          */
         "openAnnotation": boolean;
         /**
+          * A BegripResolver that will be called for STOP element "IntRef" with
+          * @scope ="Begrip".
+         */
+        "ozonContentBegripResolver"?: OzonContentBegripResolver;
+        /**
           * A UrlResolver that will be called for all STOP elements that render to HTML5 elements with external references.
          */
         "ozonContentUrlResolver"?: OzonContentUrlResolver;
@@ -1013,6 +1018,10 @@ export namespace Components {
     }
     interface DsoOzonContent {
         /**
+          * A BegripResolver that will be called for STOP element "IntRef" with scope="Begrip".
+         */
+        "begripResolver"?: OzonContentBegripResolver;
+        /**
           * The XML to be rendered.
          */
         "content"?: OzonContentInputType;
@@ -1029,6 +1038,12 @@ export namespace Components {
           * A UrlResolver that will be called for all STOP elements that render to HTML5 elements with external references.
          */
         "urlResolver"?: OzonContentUrlResolver;
+    }
+    interface DsoOzonContentToggletip {
+        /**
+          * The alias of the icon in the button.
+         */
+        "icon"?: string;
     }
     interface DsoPagination {
         /**
@@ -1387,6 +1402,11 @@ export namespace Components {
           * @default false
          */
         "openDefault": boolean;
+        /**
+          * A BegripResolver that will be called for STOP element "IntRef" with
+          * @scope ="Begrip".
+         */
+        "ozonContentBegripResolver"?: OzonContentBegripResolver1;
         /**
           * A UrlResolver that will be called for all STOP elements that render to HTML5 elements with external references.
          */
@@ -1854,7 +1874,7 @@ declare global {
         "dsoRecursiveToggle": DocumentComponentRecursiveToggleEvent;
         "dsoOpenToggle": DocumentComponentOpenToggleEvent;
         "dsoTableOfContentsClick": DocumentComponentTableOfContentsClickEvent;
-        "dsoOzonContentAnchorClick": DocumentComponentOzonContentAnchorClickEvent;
+        "dsoOzonContentClick": DocumentComponentOzonContentClickEvent;
         "dsoAnnotationToggle": DocumentComponentToggleAnnotationEvent;
         "dsoMarkItemHighlight": DocumentComponentMarkItemHighlightEvent;
     }
@@ -2197,7 +2217,6 @@ declare global {
     };
     interface HTMLDsoOzonContentElementEventMap {
         "dsoClick": OzonContentClickEvent;
-        "dsoAnchorClick": OzonContentAnchorClickEvent;
         "dsoOzonContentMarkItemHighlight": OzonContentMarkItemHighlightEvent;
     }
     interface HTMLDsoOzonContentElement extends Components.DsoOzonContent, HTMLStencilElement {
@@ -2213,6 +2232,12 @@ declare global {
     var HTMLDsoOzonContentElement: {
         prototype: HTMLDsoOzonContentElement;
         new (): HTMLDsoOzonContentElement;
+    };
+    interface HTMLDsoOzonContentToggletipElement extends Components.DsoOzonContentToggletip, HTMLStencilElement {
+    }
+    var HTMLDsoOzonContentToggletipElement: {
+        prototype: HTMLDsoOzonContentToggletipElement;
+        new (): HTMLDsoOzonContentToggletipElement;
     };
     interface HTMLDsoPaginationElementEventMap {
         "dsoSelectPage": PaginationSelectPageEvent;
@@ -2496,7 +2521,7 @@ declare global {
         new (): HTMLDsoViewerGridElement;
     };
     interface HTMLDsotDocumentComponentDemoElementEventMap {
-        "dsotOzonContentAnchorClick": DocumentComponentOzonContentAnchorClickEvent;
+        "dsotOzonContentClick": DocumentComponentOzonContentClickEvent;
         "dsotTableOfContentsClick": DocumentComponentTableOfContentsClickEvent;
     }
     interface HTMLDsotDocumentComponentDemoElement extends Components.DsotDocumentComponentDemo, HTMLStencilElement {
@@ -2558,6 +2583,7 @@ declare global {
         "dso-modal": HTMLDsoModalElement;
         "dso-onboarding-tip": HTMLDsoOnboardingTipElement;
         "dso-ozon-content": HTMLDsoOzonContentElement;
+        "dso-ozon-content-toggletip": HTMLDsoOzonContentToggletipElement;
         "dso-pagination": HTMLDsoPaginationElement;
         "dso-panel": HTMLDsoPanelElement;
         "dso-plekinfo-card": HTMLDsoPlekinfoCardElement;
@@ -3148,9 +3174,9 @@ declare namespace LocalJSX {
          */
         "onDsoOpenToggle"?: (event: DsoDocumentComponentCustomEvent<DocumentComponentOpenToggleEvent>) => void;
         /**
-          * Emitted when the user actives intRef or intIoRef anchors in Ozon Content
+          * Emitted when the user interacts with Kop, IntRef or the Kenmerken en kaart button of IntIoRef in Ozon Content
          */
-        "onDsoOzonContentAnchorClick"?: (event: DsoDocumentComponentCustomEvent<DocumentComponentOzonContentAnchorClickEvent>) => void;
+        "onDsoOzonContentClick"?: (event: DsoDocumentComponentCustomEvent<DocumentComponentOzonContentClickEvent>) => void;
         /**
           * Emitted when the user activates the recursive toggle.
          */
@@ -3169,6 +3195,11 @@ declare namespace LocalJSX {
           * @default false
          */
         "openAnnotation"?: boolean;
+        /**
+          * A BegripResolver that will be called for STOP element "IntRef" with
+          * @scope ="Begrip".
+         */
+        "ozonContentBegripResolver"?: OzonContentBegripResolver;
         /**
           * A UrlResolver that will be called for all STOP elements that render to HTML5 elements with external references.
          */
@@ -3693,6 +3724,10 @@ declare namespace LocalJSX {
     }
     interface DsoOzonContent {
         /**
+          * A BegripResolver that will be called for STOP element "IntRef" with scope="Begrip".
+         */
+        "begripResolver"?: OzonContentBegripResolver;
+        /**
           * The XML to be rendered.
          */
         "content"?: OzonContentInputType;
@@ -3706,11 +3741,7 @@ declare namespace LocalJSX {
          */
         "mark"?: OzonContentMarkFunction;
         /**
-          * Emitted when `<a>` that is created through <IntIoRef> or <IntRef> is clicked.
-         */
-        "onDsoAnchorClick"?: (event: DsoOzonContentCustomEvent<OzonContentAnchorClickEvent>) => void;
-        /**
-          * Emitted when an interactive element is clicked, except for <IntIoRef> and <IntRef>.
+          * Emitted when an interactive element is clicked.
          */
         "onDsoClick"?: (event: DsoOzonContentCustomEvent<OzonContentClickEvent>) => void;
         /**
@@ -3721,6 +3752,12 @@ declare namespace LocalJSX {
           * A UrlResolver that will be called for all STOP elements that render to HTML5 elements with external references.
          */
         "urlResolver"?: OzonContentUrlResolver;
+    }
+    interface DsoOzonContentToggletip {
+        /**
+          * The alias of the icon in the button.
+         */
+        "icon"?: string;
     }
     interface DsoPagination {
         /**
@@ -4137,9 +4174,9 @@ declare namespace LocalJSX {
          */
         "mode"?: DocumentComponentMode;
         /**
-          * To demo user interacting with IntRef or IntIoRef elements.
+          * To demo the user interacting with Kop, IntRef or the Kenmerken en kaart button of IntIoRef in Ozon Content
          */
-        "onDsotOzonContentAnchorClick"?: (event: DsotDocumentComponentDemoCustomEvent<DocumentComponentOzonContentAnchorClickEvent>) => void;
+        "onDsotOzonContentClick"?: (event: DsotDocumentComponentDemoCustomEvent<DocumentComponentOzonContentClickEvent>) => void;
         /**
           * To demo user interacting the heading in mode="table-of-contents".
          */
@@ -4149,6 +4186,11 @@ declare namespace LocalJSX {
           * @default false
          */
         "openDefault"?: boolean;
+        /**
+          * A BegripResolver that will be called for STOP element "IntRef" with
+          * @scope ="Begrip".
+         */
+        "ozonContentBegripResolver"?: OzonContentBegripResolver1;
         /**
           * A UrlResolver that will be called for all STOP elements that render to HTML5 elements with external references.
          */
@@ -4204,6 +4246,7 @@ declare namespace LocalJSX {
         "dso-modal": DsoModal;
         "dso-onboarding-tip": DsoOnboardingTip;
         "dso-ozon-content": DsoOzonContent;
+        "dso-ozon-content-toggletip": DsoOzonContentToggletip;
         "dso-pagination": DsoPagination;
         "dso-panel": DsoPanel;
         "dso-plekinfo-card": DsoPlekinfoCard;
@@ -4279,6 +4322,7 @@ declare module "@stencil/core" {
             "dso-modal": LocalJSX.DsoModal & JSXBase.HTMLAttributes<HTMLDsoModalElement>;
             "dso-onboarding-tip": LocalJSX.DsoOnboardingTip & JSXBase.HTMLAttributes<HTMLDsoOnboardingTipElement>;
             "dso-ozon-content": LocalJSX.DsoOzonContent & JSXBase.HTMLAttributes<HTMLDsoOzonContentElement>;
+            "dso-ozon-content-toggletip": LocalJSX.DsoOzonContentToggletip & JSXBase.HTMLAttributes<HTMLDsoOzonContentToggletipElement>;
             "dso-pagination": LocalJSX.DsoPagination & JSXBase.HTMLAttributes<HTMLDsoPaginationElement>;
             "dso-panel": LocalJSX.DsoPanel & JSXBase.HTMLAttributes<HTMLDsoPanelElement>;
             "dso-plekinfo-card": LocalJSX.DsoPlekinfoCard & JSXBase.HTMLAttributes<HTMLDsoPlekinfoCardElement>;