+++ packages/dso-toolkit/src/components/ozon-content/ozon-content.models.ts
@@ -3,9 +3,25 @@ export interface OzonContent {
   inline?: boolean;
   mark?: OzonContentMarkFunction;
   urlResolver?: OzonContentUrlResolver;
-  dsoAnchorClick: (e: CustomEvent) => void;
+  dsoClick: (e: CustomEvent<OzonContentClickEvent>) => void;
   dsoOzonContentMarkItemHighlight: (e: CustomEvent) => void;
+  begripResolver?: OzonContentBegripResolver;
 }
+
+interface OzonContentClickBaseEvent<T extends string> {
+  type: T;
+  node: Element;
+  originalEvent: MouseEvent;
+}
+
+export type OzonContentClickKopEvent = OzonContentClickBaseEvent<"Kop">;
+export type OzonContentClickIntIoRefEvent = OzonContentClickBaseEvent<"IntIoRef">;
+export type OzonContentClickIntRefEvent = OzonContentClickBaseEvent<"IntRef"> & { isModifiedEvent: boolean };
+
+export type OzonContentClickEvent =
+  | OzonContentClickKopEvent
+  | OzonContentClickIntRefEvent
+  | OzonContentClickIntIoRefEvent;
 
 export type OzonContentMarkFunction = (text: string) => OzonContentText[] | undefined;
 
@@ -18,5 +34,14 @@ export interface OzonContentMarkItem {
 
 export interface OzonContentUrlResolver {
   (name: "Illustratie" | "InlineTekstAfbeelding", attribute: "naam", value: string | null, element: Element): string;
-  (name: "ExtIoRef" | "ExtRef", attribute: "ref", value: string | null, element: Element): string;
+  (
+    name: "ExtIoRef" | "ExtRef" | "IntIoRef" | "IntRef",
+    attribute: "ref",
+    value: string | null,
+    element: Element,
+  ): string;
+}
+
+export interface OzonContentBegripResolver {
+  (ref: string | null, element: Element): XMLDocument | string | undefined;
 }